Added necessary tweak
This adds a tweak to change the Big Picture Mode icon to the controller
icon. Currently, in the override folder, there includes a replacement
steam.styles file. I have removed the replacement steam.styles file (as
this is something that should never have to be replaced) and added a
tweak for the button change instead.